require non null, remove abs in assert, move rescaler to another package
