+++ flow-decomposition/src/main/java/com/powsybl/flow_decomposition/rescaler/DecomposedFlowRescalerAcerMethodology.java
@@ -5,8 +5,10 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.flow_decomposition;
+package com.powsybl.flow_decomposition.rescaler;
 
+import com.powsybl.flow_decomposition.DecomposedFlow;
+import com.powsybl.flow_decomposition.DecomposedFlowBuilder;
 import com.powsybl.iidm.network.Country;
 
 import java.util.Map;
@@ -17,24 +19,24 @@ import java.util.stream.Collectors;
  * @author Hugo Schindler {@literal <hugo.schindler at rte-france.com>}
  * @author Caio Luke {@literal <caio.luke at artelys.com>}
  */
-public class DecomposedFlowRescalerProportional implements DecomposedFlowRescaler {
-    public static final double MIN_FLOW_TOLERANCE = 1E-6; // min flow in MW to rescale
+public class DecomposedFlowRescalerAcerMethodology implements DecomposedFlowRescaler {
 
-    public DecomposedFlowRescalerProportional() {
+    public DecomposedFlowRescalerAcerMethodology() {
         // empty constructor
+    }
+
+    private static double reLU(double value) {
+        return value > 0 ? value : 0.;
+    }
+
+    private static double rescaleValue(double initialValue, double delta, double sumOfReLUFlows) {
+        return initialValue + delta * reLU(initialValue) / sumOfReLUFlows;
     }
 
     @Override
     public DecomposedFlow rescale(DecomposedFlow decomposedFlow) {
         double acTerminal1ReferenceFlow = decomposedFlow.getAcTerminal1ReferenceFlow();
-        double acTerminal2ReferenceFlow = decomposedFlow.getAcTerminal2ReferenceFlow();
-        if (Double.isNaN(acTerminal1ReferenceFlow) || Double.isNaN(acTerminal2ReferenceFlow)) {
-            return decomposedFlow;
-        }
-
-        // if dcReferenceFlow is too small, do not rescale
-        double dcReferenceFlow = decomposedFlow.getDcReferenceFlow();
-        if (Math.abs(dcReferenceFlow) < MIN_FLOW_TOLERANCE) {
+        if (Double.isNaN(acTerminal1ReferenceFlow)) {
             return decomposedFlow;
         }
 
@@ -42,22 +44,23 @@ public class DecomposedFlowRescalerProportional implements DecomposedFlowRescale
         String contingencyId = decomposedFlow.getContingencyId();
         Country country1 = decomposedFlow.getCountry1();
         Country country2 = decomposedFlow.getCountry2();
+        double acTerminal2ReferenceFlow = decomposedFlow.getAcTerminal2ReferenceFlow();
+        double dcReferenceFlow = decomposedFlow.getDcReferenceFlow();
         double allocatedFlow = decomposedFlow.getAllocatedFlow();
         double xNodeFlow = decomposedFlow.getXNodeFlow();
         double pstFlow = decomposedFlow.getPstFlow();
         double internalFlow = decomposedFlow.getInternalFlow();
         Map<String, Double> loopFlows = decomposedFlow.getLoopFlows();
 
-        // rescale proportionally to max (abs) ac flow
-        double acMaxAbsFlow = decomposedFlow.getMaxAbsAcFlow();
-        double rescaleFactor = Math.abs(acMaxAbsFlow / dcReferenceFlow);
+        double deltaToRescale = acTerminal1ReferenceFlow * Math.signum(acTerminal1ReferenceFlow) - decomposedFlow.getTotalFlow();
+        double sumOfReLUFlows = reLU(allocatedFlow) + reLU(pstFlow) + reLU(xNodeFlow) + loopFlows.values().stream().mapToDouble(DecomposedFlowRescalerAcerMethodology::reLU).sum() + reLU(internalFlow);
 
-        double rescaledAllocatedFlow = rescaleFactor * allocatedFlow;
-        double rescaledXNodeFlow = rescaleFactor * xNodeFlow;
-        double rescaledPstFlow = rescaleFactor * pstFlow;
-        double rescaleInternalFlow = rescaleFactor * internalFlow;
+        double rescaledAllocatedFlow = rescaleValue(allocatedFlow, deltaToRescale, sumOfReLUFlows);
+        double rescaledXNodeFlow = rescaleValue(xNodeFlow, deltaToRescale, sumOfReLUFlows);
+        double rescaledPstFlow = rescaleValue(pstFlow, deltaToRescale, sumOfReLUFlows);
+        double rescaleInternalFlow = rescaleValue(internalFlow, deltaToRescale, sumOfReLUFlows);
         Map<String, Double> rescaledLoopFlows = loopFlows.entrySet().stream()
-                .collect(Collectors.toMap(Map.Entry::getKey, entry -> rescaleFactor * entry.getValue()));
+                .collect(Collectors.toMap(Map.Entry::getKey, entry -> rescaleValue(entry.getValue(), deltaToRescale, sumOfReLUFlows)));
 
         return new DecomposedFlowBuilder()
                 .addBranchId(branchId)
