+++ flow-decomposition/src/test/java/com/powsybl/flow_decomposition/InternalFlowTests.java
@@ -6,6 +6,7 @@
  */
 package com.powsybl.flow_decomposition;
 
+import com.powsybl.flow_decomposition.rescaler.DecomposedFlowRescalerAcerMethodology;
 import com.powsybl.flow_decomposition.xnec_provider.XnecProviderAllBranches;
 import com.powsybl.iidm.network.Country;
 import com.powsybl.iidm.network.Network;
