+++ flow-decomposition/src/test/java/com/powsybl/flow_decomposition/TestUtils.java
@@ -41,9 +41,9 @@ public final class TestUtils {
         for (String xnec : flowDecompositionResults.getDecomposedFlowMap().keySet()) {
             DecomposedFlow decomposedFlow = flowDecompositionResults.getDecomposedFlowMap().get(xnec);
             switch (rescaleMode) {
-                case ACER_METHODOLOGY -> assertEquals(Math.abs(decomposedFlow.getAcTerminal1ReferenceFlow()), Math.abs(decomposedFlow.getTotalFlow()), EPSILON);
-                case PROPORTIONAL -> assertEquals(decomposedFlow.getMaxAbsAcFlow(), Math.abs(decomposedFlow.getTotalFlow()), EPSILON);
-                default -> assertEquals(Math.abs(decomposedFlow.getDcReferenceFlow()), Math.abs(decomposedFlow.getTotalFlow()), EPSILON);
+                case ACER_METHODOLOGY -> assertEquals(Math.abs(decomposedFlow.getAcTerminal1ReferenceFlow()), decomposedFlow.getTotalFlow(), EPSILON);
+                case PROPORTIONAL -> assertEquals(decomposedFlow.getMaxAbsAcFlow(), decomposedFlow.getTotalFlow(), EPSILON);
+                default -> assertEquals(Math.abs(decomposedFlow.getDcReferenceFlow()), decomposedFlow.getTotalFlow(), EPSILON);
             }
         }
     }
